sys/console: replaced old console implementation with new
Deleted minimal console implementation, changed console API
to match the new console API. Some things are WIP or TODO.
