+++ sys/log/full/src/log_console.c
@@ -29,18 +29,19 @@
 static int
 log_console_append(struct log *log, void *buf, int len)
 {
-    struct log_entry_hdr *hdr;
+//    struct log_entry_hdr *hdr;
 
     if (!console_is_init()) {
         return (0);
     }
-
-    if (!console_is_midline) {
-        hdr = (struct log_entry_hdr *) buf;
-        console_printf("[ts=%lussb, mod=%u level=%u] ",
-                (unsigned long) hdr->ue_ts, hdr->ue_module,
-                hdr->ue_level);
-    }
+    // TODO
+//
+//    if (!console_is_midline) {
+//        hdr = (struct log_entry_hdr *) buf;
+//        console_printf("[ts=%lussb, mod=%u level=%u] ",
+//                (unsigned long) hdr->ue_ts, hdr->ue_module,
+//                hdr->ue_level);
+//    }
 
     console_write((char *) buf + LOG_ENTRY_HDR_SIZE, len - LOG_ENTRY_HDR_SIZE);
