+++ sys/newshell/src/shell_os.c
@@ -24,7 +24,7 @@
 #include "datetime/datetime.h"
 #include "syscfg/syscfg.h"
 
-#include "newshell/console/console.h"
+#include "console/console.h"
 #include "newshell/shell/shell.h"
 
 #define SHELL_OS "os"
