+++ sys/newshell/syscfg.yml
@@ -40,17 +40,3 @@ syscfg.defs:
     SHELL_PROMPT_MODULE:
         description: 'Include shell prompt module'
         value: 1
-
-    CONSOLE_RTT:
-        description: 'Enable console over RTT'
-        value: 0
-    CONSOLE_UART:
-        description: 'Enable console over UART'
-        value: 0
-
-    CONSOLE_UART_BAUD:
-        description: 'Console UART baud rate.'
-        value: '115200'
-    CONSOLE_UART_FLOW_CONTROL:
-        description: 'Console UART flow control.'
-        value: 'UART_FLOW_CTL_NONE'
